Implement TagCommand
Added the TagCommand class to allow users to add or
modify tags for a person in the address book. The
TagCommandParser has been implemented to parse user
input for tagging, accepting a person's index or
identifier and one or more tags. This command provides
more flexibility for managing person tags.

- Implemented TagCommand class
- Created TagCommandParser to handle user input
- Updated command usage messages and help
- Tested the TagCommand and TagCommandParser